refactor: small adjustments for Tour Review

+++ src/Modules/Tours/Explorer.Tours.API/Dtos/TourReviewDto.cs
@@ -9,11 +9,11 @@ namespace Explorer.Tours.API.Dtos;
 public class TourReviewDto
 {
     public long Id { get; set; }
-    public int TourId { get; set; }
-    public int ExecutionId { get; set; }
+    public long TourId { get; set; }
+    public long ExecutionId { get; set; }
     public int Rating { get; set; }
     public string Comment { get; set; }
-    public int TouristId { get; set; }
+    public long TouristId { get; set; }
     public DateTime ReviewDate { get; set; }
     public float CompletedPercentage { get; set; }
     public List<string> Images { get; set; }
