+++ src/Modules/Tours/Explorer.Tours.API/Public/Administration/ITourReviewService.cs
@@ -6,9 +6,9 @@ namespace Explorer.Tours.API.Public.Administration;
 public interface ITourReviewService
 {
     PagedResult<TourReviewDto> GetPaged(int page, int pageSize);
-    TourReviewDto Create(TourReviewDto review, int touristId);
-    TourReviewDto Update(TourReviewDto review, int touristId);
-    public void Delete(long id, int touristId);
-    double GetAverageRating(int tourId);
-    PagedResult<TourReviewDto> GetByTourId(int page, int pageSize, int tourId);
+    TourReviewDto Create(TourReviewDto review, long touristId);
+    TourReviewDto Update(TourReviewDto review, long touristId);
+    public void Delete(long id, long touristId);
+    double GetAverageRating(long tourId);
+    PagedResult<TourReviewDto> GetByTourId(int page, int pageSize, long tourId);
 }
